requirements.txt
requests
praw
beautifulsoup4
python-dotenv
boto3
